+++ frontend/src/store/modules/checks.js
@@ -1,4 +1,4 @@
-import { apiService2 } from '@/common/apiService2';
+import { apiService } from '@/common/apiService';
 
 export default {
   namespaced: true,
@@ -20,7 +20,7 @@ export default {
     async getHealthCheckStatus(context) {
       context.commit('setHealthCheck', null);
       try {
-        const response = await apiService2.getHealthCheck();
+        const response = await apiService.getHealthCheck();
         context.commit('setHealthCheck', response);
       } catch (e) {
         context.commit('setHealthCheck', 'error');
@@ -29,7 +29,7 @@ export default {
     async getApiCheck(context, route) {
       context.commit('setApiCheckResponse', '');
       try {
-        const response = await apiService2.getApiCheck(route);
+        const response = await apiService.getApiCheck(route);
         context.commit('setApiCheckResponse', response);
       } catch (e) {
         context.commit('setApiCheckResponse', e);
