+++ frontend/tests/unit/common/apiService.spec.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { apiService2 } from '@/common/apiService2';
+import { apiService } from '@/common/apiService';
 
 import * as healthCheckFixture from './fixtures/healthCheck.json';
 import * as apiCheckFixture from './fixtures/apiCheck.json';
@@ -9,7 +9,7 @@ jest.mock('axios');
 describe('getAPICheck()', () => {
   it('calls `getAPICheck() endpoint with a route', async () => {
     axios.get.mockResolvedValueOnce(apiCheckFixture.response);
-    const res = await apiService2.getApiCheck(apiCheckFixture.route);
+    const res = await apiService.getApiCheck(apiCheckFixture.route);
 
     expect(res).toContain(`URL: ${apiCheckFixture.route}`);
     expect(axios.get).toHaveBeenCalledWith(apiCheckFixture.route);
@@ -20,7 +20,7 @@ describe('getAPICheck()', () => {
 describe('getHealthCheck()', () => {
   it('calls `getHealthCheck() endpoint', async () => {
     axios.get.mockResolvedValueOnce({ data: healthCheckFixture });
-    const check = await apiService2.getHealthCheck();
+    const check = await apiService.getHealthCheck();
 
     expect(check).toEqual(healthCheckFixture);
     expect(axios.get).toHaveBeenCalledWith('/api/v1/checks/status');
